+++ components/servicedesk-plus/src/actions/products/create.ts
@@ -5,10 +5,8 @@ import {
   attributes,
   connectionInput,
   id,
-  isLaptop,
-  manufacturer,
   name,
-  partNo,
+  productDetails,
   productType,
 } from "../../inputs";
 import { createPayload } from "../../util";
@@ -27,24 +25,13 @@ export const createProduct = action({
       required: true,
     },
     id,
-    manufacturer,
-    isLaptop,
-    partNo,
+    productDetails,
     attributes,
     connectionInput,
   },
   perform: async (
     context,
-    {
-      connectionInput,
-      id,
-      attributes,
-      name,
-      productType,
-      isLaptop,
-      manufacturer,
-      partNo,
-    },
+    { connectionInput, id, attributes, name, productType, productDetails },
   ) => {
     const client = createClient(connectionInput, context.debug.enabled);
     const payload = createPayload({
@@ -52,9 +39,9 @@ export const createProduct = action({
         id,
         name,
         product_type: productType,
-        is_laptop: isLaptop,
-        manufacturer,
-        parto_no: partNo,
+        is_laptop: productDetails.isLaptop,
+        manufacturer: productDetails.manufacturer,
+        parto_no: productDetails.partNo,
         ...attributes,
       },
     });